requirements.txt
fastapi>=0.68.0,<0.109.0
uvicorn>=0.15.0
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pymysql
sqlalchemy>=1.4.0,<2.0.0
pydantic>=1.10.0,<2.0.0
email-validator>=1.3.1
python-dateutil>=2.8.2